+++ template/method_data.go
@@ -2,100 +2,8 @@ package template
 
 import (
 	"fmt"
-	"go/types"
 	"strings"
 )
-
-type Packages []*Package
-
-func (p Packages) SyncPkgQualifier() string {
-	for _, imprt := range p {
-		if imprt.Path() == "sync" {
-			return imprt.Qualifier()
-		}
-	}
-
-	return "sync"
-}
-
-// Data is the template data used to render the mock template.
-type Data struct {
-	PkgName         string
-	SrcPkgQualifier string
-	Imports         []*Package
-	Mocks           []MockData
-	TemplateData    map[string]any
-}
-
-// MocksSomeMethod returns true of any one of the Mocks has at least 1
-// method.
-func (d Data) MocksSomeMethod() bool {
-	for _, m := range d.Mocks {
-		if len(m.Methods) > 0 {
-			return true
-		}
-	}
-
-	return false
-}
-
-// MockData is the data used to generate a mock for some interface.
-type MockData struct {
-	InterfaceName string
-	MockName      string
-	TypeParams    []TypeParamData
-	Methods       []MethodData
-	TemplateData  map[string]any
-}
-
-func (m MockData) TypeConstraintTest() string {
-	if len(m.TypeParams) == 0 {
-		return ""
-	}
-	s := "["
-	for idx, param := range m.TypeParams {
-		if idx != 0 {
-			s += ", "
-		}
-		s += exported(param.Name())
-		s += " "
-		s += param.TypeString()
-	}
-	s += "]"
-	return s
-}
-
-func (m MockData) TypeConstraint() string {
-	if len(m.TypeParams) == 0 {
-		return ""
-	}
-	s := "["
-	for idx, param := range m.TypeParams {
-		if idx != 0 {
-			s += ", "
-		}
-		s += exported(param.Name())
-		s += " "
-		s += param.TypeString()
-	}
-	s += "]"
-	return s
-}
-
-func (m MockData) TypeInstantiation() string {
-	if len(m.TypeParams) == 0 {
-		return ""
-	}
-	s := "["
-	for idx, param := range m.TypeParams {
-		if idx != 0 {
-			s += ", "
-		}
-		s += exported(param.Name())
-	}
-	s += "]"
-	return s
-}
 
 // MethodData is the data which represents a method on some interface.
 type MethodData struct {
@@ -281,67 +189,4 @@ func (m MethodData) ReturnArgList() string {
 
 func (m MethodData) IsVariadic() bool {
 	return len(m.Params) > 0 && m.Params[len(m.Params)-1].Variadic
-}
-
-type TypeParamData struct {
-	ParamData
-	Constraint types.Type
-}
-
-// ParamData is the data which represents a parameter to some method of
-// an interface.
-type ParamData struct {
-	Var      *Var
-	Variadic bool
-}
-
-// Name returns the name of the parameter.
-func (p ParamData) Name() string {
-	return p.Var.Name
-}
-
-// MethodArg is the representation of the parameter in the function
-// signature, ex: 'name a.Type'.
-func (p ParamData) MethodArg() string {
-	if p.Variadic {
-		return fmt.Sprintf("%s ...%s", p.Name(), p.TypeString()[2:])
-	}
-	return fmt.Sprintf("%s %s", p.Name(), p.TypeString())
-}
-
-// CallName returns the string representation of the parameter to be
-// used for a method call. For a variadic paramter, it will be of the
-// format 'foos...' if ellipsis is true.
-func (p ParamData) CallName(ellipsis bool) string {
-	if ellipsis && p.Variadic {
-		return p.Name() + "..."
-	}
-	return p.Name()
-}
-
-// TypeString returns the string representation of the type of the
-// parameter.
-func (p ParamData) TypeString() string {
-	return p.Var.TypeString()
-}
-
-// TypeStringEllipsis returns the string representation of the type of the
-// parameter. If it is a variadic parameter, it will be represented as a
-// variadic parameter instead of a slice. For example instead of `[]string`,
-// it will return `...string`.
-func (p ParamData) TypeStringEllipsis() string {
-	typeString := p.TypeString()
-	if !p.Variadic {
-		return typeString
-	}
-	return strings.Replace(typeString, "[]", "...", 1)
-}
-
-// TypeStringVariadicUnderlying returns the underlying type of a variadic parameter. For
-// instance, if a function has a parameter defined as `foo ...int`, this function
-// will return "int". If the parameter is not variadic, this will behave the same
-// as `TypeString`.
-func (p ParamData) TypeStringVariadicUnderlying() string {
-	typeString := p.TypeStringEllipsis()
-	return strings.Replace(typeString, "...", "", 1)
 }
