+++ template/package.go
@@ -1,9 +1,5 @@
 package template
 
-import (
-	"strings"
-)
-
 type TypesPackage interface {
 	Name() string
 	Path() string
@@ -49,38 +45,4 @@ func (p *Package) Path() string {
 	}
 
 	return p.pkg.Path()
-}
-
-var replacer = strings.NewReplacer(
-	"go-", "",
-	"-go", "",
-	"-", "",
-	"_", "",
-	".", "",
-	"@", "",
-	"+", "",
-	"~", "",
-)
-
-// uniqueName generates a unique name for a package by concatenating
-// path components. The generated name is guaranteed to unique with an
-// appropriate level because the full package import paths themselves
-// are unique.
-func (p Package) uniqueName(lvl int) string {
-	pp := strings.Split(p.Path(), "/")
-	reverse(pp)
-
-	var name string
-	for i := 0; i < min(len(pp), lvl+1); i++ {
-		name = strings.ToLower(replacer.Replace(pp[i])) + name
-	}
-
-	return name
-}
-
-func reverse(a []string) {
-	for i := len(a)/2 - 1; i >= 0; i-- {
-		opp := len(a) - 1 - i
-		a[i], a[opp] = a[opp], a[i]
-	}
 }
