+++ template/packages.go
@@ -1,0 +1,13 @@
+package template
+
+type Packages []*Package
+
+func (p Packages) SyncPkgQualifier() string {
+	for _, imprt := range p {
+		if imprt.Path() == "sync" {
+			return imprt.Qualifier()
+		}
+	}
+
+	return "sync"
+}
