+++ template/type_param_data.go
@@ -1,0 +1,8 @@
+package template
+
+import "go/types"
+
+type TypeParamData struct {
+	ParamData
+	Constraint types.Type
+}
